+++ RiskOfTheAncients2/Equipment/BlackKingBar.cs
@@ -2,8 +2,6 @@
 using ROTA2.Buffs;
 using RoR2;
 using UnityEngine;
-using IL.RoR2.WwiseUtils;
-using RoR2.WwiseUtils;
 
 namespace ROTA2.Equipment
 {
@@ -33,11 +31,10 @@ namespace ROTA2.Equipment
             BlackKingBarCooldown = config.Bind("Equipment: " + EquipmentName, "Cooldown",                  75.0f, "").Value;
         }
 
-        public uint SoundID = 3471834074;
         protected override bool ActivateEquipment(EquipmentSlot slot)
         {
             var body = slot.characterBody;
-            if (body)
+            if (HasThisEquipment(body))
             {
                 BlackKingBarBuff.Instance.ApplyTo(new BuffBase.ApplyParameters
                 {